+++ packages/cloud-cognitive/src/components/Datagrid/Datagrid/DatagridToolbar.js
@@ -161,6 +161,8 @@ const DatagridToolbar = (datagridState) => {
     datagridState;
   const { filterTags, EventEmitter } = useContext(FilterContext);
 
+  console.log({ filters: state.filters });
+
   const renderFilterSummary = () =>
     state.filters.length > 0 && (
       <FilterSummary
